Module Wodan_irmin

module Log : Logs.LOG

val standard_mount_options : Wodan.mount_options

module StandardSuperblockParams : Wodan.SUPERBLOCK_PARAMS

Defaults for SUPERBLOCK_PARAMS

module Conf : sig ... end

val config : ?⁠config:Irmin.config -> path:string -> create:bool -> ?⁠cache_size:int -> ?⁠fast_scan:bool -> ?⁠list_key:string -> ?⁠autoflush:bool -> unit -> Irmin.config

module type BLOCK_CON = sig ... end

module type DB = sig ... end

module DB_BUILDER : functor (_ : BLOCK_CON) -> functor (_ : Wodan.SUPERBLOCK_PARAMS) -> DB

Builds a Wodan-backed database, that can be used to build higher-level Irmin interfaces

module AO_BUILDER : functor (_ : DB) -> Irmin.APPEND_ONLY_STORE_MAKER

Builds an Irmin.APPEND_ONLY_STORE, storing key to value mappings

module CA_BUILDER : functor (_ : DB) -> Irmin.CONTENT_ADDRESSABLE_STORE_MAKER

Builds an Irmin.CONTENT_ADDRESSABLE_STORE, storing values through stable hashing

module AW_BUILDER : functor (_ : DB) -> functor (_ : Irmin.Hash.S) -> Irmin.ATOMIC_WRITE_STORE_MAKER

Builds an Irmin.ATOMIC_WRITE_STORE, storing key to value mappings with extra features

module Make : functor (DB : DB) -> functor (M : Irmin.Metadata.S) -> functor (C : Irmin.Contents.S) -> functor (P : Irmin.Path.S) -> functor (B : Irmin.Branch.S) -> functor (H : Irmin.Hash.S) -> sig ... end

module Make_chunked : functor (DB : DB) -> functor (M : Irmin.Metadata.S) -> functor (C : Irmin.Contents.S) -> functor (P : Irmin.Path.S) -> functor (B : Irmin.Branch.S) -> functor (H : Irmin.Hash.S) -> sig ... end

module KV : functor (DB : DB) -> functor (H : Irmin.Hash.S) -> functor (C : Irmin.Contents.S) -> sig ... end

module KV_git : functor (DB : DB) -> functor (H : Irmin.Hash.S) -> sig ... end

module KV_git_sha1 : functor (DB : DB) -> sig ... end

module KV_chunked : functor (DB : DB) -> functor (H : Irmin.Hash.S) -> functor (C : Irmin.Contents.S) -> sig ... end